Refactor: Move categories-related logic to categories.py
- Created a new file  under the  directory.
- Moved all categories-related endpoints (, , ) from  to .
- Used  for modular route handling.
- Updated  to include the  with the prefix .
- Improved code organization and maintainability by separating route logic.
