+++ app/main.py
@@ -3,8 +3,9 @@ from fastapi import FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from contextlib import asynccontextmanager
 from app.database import get_database
-from app.models import postCommunityPost, postedCategory, postCategory, deleteCategory, postJoke, user
+from app.models import postCommunityPost, postJoke, user
 from app.routes.community_posts import router as community_posts_router
+from app.routes.categories import router as categories_router
 
 from dotenv import load_dotenv
 from bson import ObjectId
@@ -30,7 +31,7 @@ origins = [
 ]
 
 app.include_router(community_posts_router, prefix="/community-posts", tags=["Community Posts"])
-
+app.include_router(categories_router, prefix="/categories", tags=["Categories"])
 
 app.add_middleware(
     CORSMiddleware,
@@ -44,50 +45,65 @@ app.add_middleware(
 async def root():
     return {"message": "FastAPI with MongoDB"}
 
+# @app.get("/community-posts")
+# async def get_community_posts():
+#     posts = await db["communityPost"].find().to_list(length=100)
+#     for post in posts:
+#         post["id"] = str(post["_id"])
+#         del post["_id"]
+#     return posts
+
+# @app.post("/post-community-post")
+# async def post_community_post(post: postCommunityPost):
+#     post_dict = post.model_dump()
+#     result = await db["communityPost"].insert_one(post_dict)
+#     if result.inserted_id:
+#         return {"message": "Post Created Successfully"}
+#     else:
+#         raise HTTPException(status_code=500, detail="Failed to create post")
+
+# @app.get("/posted-categories", response_model=list[postedCategory])
+# async def posted_categories():
+#     postedCategories = await db["postedCategory"].find().to_list(length=100)
+#     for category in postedCategories:
+#         category["id"] = str(category["_id"])
+#         del category["_id"]
+#         # Handle timestamp conversion
+#         if isinstance(category["timestamp"], int):  # If timestamp is an integer
+#             category["timestamp"] = datetime.fromtimestamp(category["timestamp"] / 1000).isoformat()
+#         elif isinstance(category["timestamp"], datetime):  # If timestamp is a datetime object
+#             category["timestamp"] = category["timestamp"].isoformat()
+#     return postedCategories
 
 
-@app.get("/posted-categories", response_model=list[postedCategory])
-async def posted_categories():
-    postedCategories = await db["postedCategory"].find().to_list(length=100)
-    for category in postedCategories:
-        category["id"] = str(category["_id"])
-        del category["_id"]
-        # Handle timestamp conversion
-        if isinstance(category["timestamp"], int):  # If timestamp is an integer
-            category["timestamp"] = datetime.fromtimestamp(category["timestamp"] / 1000).isoformat()
-        elif isinstance(category["timestamp"], datetime):  # If timestamp is a datetime object
-            category["timestamp"] = category["timestamp"].isoformat()
-    return postedCategories
+
+# @app.post("/add-categoryPost")
+# async def post_category(category: postCategory):
+#     category_dict = category.model_dump()
+#     result = await db["postedCategory"].insert_one(category_dict)
+#     if result.inserted_id:
+#         return {"message": "Category Posted Successfully"}
+#     else:
+#         raise HTTPException(status_code=500, detail="Failed to post category")
 
 
 
-@app.post("/add-categoryPost")
-async def post_category(category: postCategory):
-    category_dict = category.model_dump()
-    result = await db["postedCategory"].insert_one(category_dict)
-    if result.inserted_id:
-        return {"message": "Category Posted Successfully"}
-    else:
-        raise HTTPException(status_code=500, detail="Failed to post category")
+# @app.delete("/delete-posted-category")
+# async def delete_posted_category(request: deleteCategory):
+#     try:
+#         print('Received ID:', request.id)
+#         # Convert the string id to ObjectId
+#         object_id = ObjectId(request.id)
+#         print('ObjectId:', object_id)
+#     except Exception as e:
+#         print("Error occurred:", str(e))
+#         raise HTTPException(status_code=400, detail="Invalid category ID format")
 
-
-
-@app.delete("/delete-posted-category")
-async def delete_posted_category(request: deleteCategory):
-    try:
-        print('Received ID:', request.id)
-        # Convert the string id to ObjectId
-        object_id = ObjectId(request.id)
-        print('ObjectId:', object_id)
-    except Exception as e:
-        print("Error occurred:", str(e))
-        raise HTTPException(status_code=400, detail="Invalid category ID format")
-
-    result = await db["postedCategory"].delete_one({"_id": object_id})
-    if result.deleted_count == 1:
-        return {"message": "Category deleted successfully"}
-    else:
-        raise HTTPException(status_code=404, detail="Category not found")
+#     result = await db["postedCategory"].delete_one({"_id": object_id})
+#     if result.deleted_count == 1:
+#         return {"message": "Category deleted successfully"}
+#     else:
+#         raise HTTPException(status_code=404, detail="Category not found")
      
 
 @app.post("/add-joke")
